initial convert debugging complete
--HG--
extra : convert_revision : svn%3A8d6eca06-2316-de11-9fcd-000c29fc710c/trunk%4026

+++ convert/include/itloader.h
@@ -104,7 +104,7 @@ namespace ITLoader {
 		u8	DefaultPan;
 		u8	RandomVolume;
 		u8	RandomPanning;
-		u8	TrackerVersion;
+		u16	TrackerVersion;
 		u8	NumberOfSamples;
 		u8	InitialFilterCutoff;
 		u8	InitialFilterResonance;
